Organize and remove redundant tests
