+++ src/test/scala/io/qbeast/spark/index/IndexTest.scala
@@ -236,4 +236,29 @@ class IndexTest
       }
     }
 
+  it should "index correctly when a small cubeSize is given" in withSpark { spark =>
+    withOTreeAlgorithm { oTreeAlgorithm =>
+      val rdd =
+        spark.sparkContext.parallelize(
+          0.to(1000)
+            .map(i => Client3(i * i, s"student-$i", i, i * 1000 + 123, i * 2567.3432143)))
+
+      val df = spark.createDataFrame(rdd)
+      val smallCubeSize = 10
+      val rev = SparkRevisionFactory.createNewRevision(
+        QTableID("test"),
+        df.schema,
+        Map("columnsToIndex" -> "age,val2", "cubeSize" -> smallCubeSize.toString))
+
+      val (indexed, tc) = oTreeAlgorithm.index(df, IndexStatus(rev))
+      val weightMap = tc.cubeWeights
+
+      checkDFSize(indexed, df)
+      checkCubes(weightMap)
+      checkWeightsIncrement(weightMap)
+      checkCubesOnData(weightMap, indexed, 2)
+      checkCubeSize(tc, rev, indexed)
+    }
+  }
+
 }
